+++ RELEASE_NOTES.md
@@ -2,6 +2,9 @@
 
 ## New in git-machete 3.10.0
 
+- added: boolean git config key `machete.status.extraSpaceBeforeBranchName` that enable configurable rendering of `status` command
+- added: 3 git config keys `machete.github.{remote,organization,repository}` that enable `git machete github *` subcommands to work with custom GitHub URLs
+
 ## New in git-machete 3.9.1
 
 - fixed: better rendering of edge junctions in `status`